+++ Cargo.toml
@@ -17,10 +17,10 @@ maintenance = { status = "actively-developed" }
 [dependencies]
 cargo-lock = { version = "4", default-features = false }
 chrono = { version = "0.4", features = ["serde"] }
-crates-index = "=0.15.2"
+crates-index = { version = "0.15.2", optional = true }
 cvss = { version = "1", features = ["serde"] }
-git2 = "0.13"
-home = "0.5"
+git2 = { version = "0.13", optional = true }
+home = { version = "0.5", optional = true }
 platforms = { version = "1", features = ["serde"] }
 semver = { version = "0.9", features = ["serde"] }
 semver-parser = "0.9"
@@ -39,6 +39,8 @@ tempfile = "3"
 once_cell = "1.3.0"
 
 [features]
+default = ["fetch"]
+fetch = ["crates-index", "git2", "home"]
 fix = ["cargo-edit"]
 dependency-tree = ["cargo-lock/dependency-tree"]
 vendored-openssl = ["git2/vendored-openssl"]
