+++ tests/database.rs
@@ -1,13 +1,20 @@
+#![cfg(feature = "fetch")]
+
 use cargo_lock::Lockfile;
 use once_cell::sync::Lazy;
 use rustsec::database::scope;
 use rustsec::database::Query;
+use rustsec::repository::git::GitRepository;
 use rustsec::Database;
 use std::path::Path;
 use std::sync::Mutex;
 
-static DEFAULT_DATABASE: Lazy<Mutex<Database>> =
-    Lazy::new(|| Mutex::new(Database::fetch().expect("Should be fetchable.")));
+static DEFAULT_DATABASE: Lazy<Mutex<Database>> = Lazy::new(|| {
+    Mutex::new(
+        Database::load(&GitRepository::fetch_default_repo().unwrap())
+            .expect("Should be fetchable."),
+    )
+});
 
 /// Queries vulnerabilites in public package scope
 #[test]
